# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nebula-mgmt API
  version: "1.0.0"
  description: >
    Control-plane REST API consumed by the CLI, the nebula-agent poller, and
    mobile bundles. Agent endpoints (enroll, agent/updates) are unauthenticated
    at the bearer layer — enroll carries a single-use token, polls carry an
    Ed25519 proof-of-possession signature (ADR 0004). Admin endpoints require an
    operator API key.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/networks/{id}:
    get:
      operationId: getNetwork
      summary: Get a network by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Operator API key as a bearer token.
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Network:
      type: object
      required: [id, name, cidrs, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        cidrs:
          type: array
          items:
            type: string
        ca_id:
          type: string
        created_at:
          type: string
          format: date-time
...